% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{MRIE_EM}
\alias{MRIE_EM}
\title{Mendelian Randomization (MR) Analysis with Incomplete Measurements on the Exposure}
\usage{
MRIE_EM(IV.dat, pheno.dat, covar.dat = NULL, epsilon = 1e-05)
}
\arguments{
\item{IV.dat}{A required data frame containing data on the instrumental variables (IVs) for all
the subjects, with a header line indicating the names of the IVs. Each row corresponds
to one subject, and each column corresponds to one IV, such as SNP dosages and
polygenic risk scores. Missing value or \code{NA} is not allowed.}

\item{pheno.dat}{A required data frame containing data on the outcome and the exposure, with a
header line indicating their names. Missing value or \code{NA} is not allowed. Each row
corresponds to one subject, and the subjects should be listed in the same order as
in \code{IV.dat}. The dataset \code{pheno.dat} should have three columns. The first column
contains the floating-point values for the continuous outcome. The second column
(integer-valued) and the third column (floating-point valued) contain information on the exposure
variable. See the "Details" section for instructions on how the second and the third column
should be specified.}

\item{covar.dat}{An optional data frame containing data on the measured covariates
(e.g., age, gender, and race), with a header line indicating the names of these variables.
Each row corresponds to one subject, and the subjects should be listed in the same order
as in \code{IV.dat}. Each column corresponds to one covariate. Missing value or \code{NA}
is not allowed.}

\item{epsilon}{The convergence threshold of the EM algorithm. The iterations will be
terminated if the Euclidean distance between the parameter values at two successive
iterations is less than this value. Default as \eqn{10^{-5}}.}
}
\value{
A list will be returned with the following components:
\itemize{
\item{\code{results_reg} contains the parameter estimates for \eqn{\gamma}
and the components in \eqn{\alpha} and \eqn{\beta}. The standard error
estimates of the estimated parameters and the corresponding p-values are also
provided. The inference on the causal effect of interest is presented in the
last row.}
\item{\code{results_var} contains the estimates for the variance components
(i.e., the variances of \eqn{\epsilonS} and \eqn{\epsilonY} and the
correlation between \eqn{\epsilonS} and \eqn{\epsilonY}).}
\item{\code{full_cov_mat} is the estimated covariance matrix of the parameter
estimates derived from the Louis formula.}
}
}
\description{
This function is used for MR analysis with a continuous outcome and a continuous
exposure that is potentially unmeasured or non-detectable. The estimations are
carried out based on the maximum likelihood estimation method, and the EM algorithm
is used for the computation. The covariance matrix of the estimated parameters is
derived from the Louis formula (Little and Rubin, 2019).
}
\details{
Let \eqn{Y} be a continuous outcome, \eqn{S} be a continuous exposure that is
potentially unmeasured or subject to detection limits, \eqn{G} be a
vector of IVs for \eqn{S}, and \eqn{Z} be a vector of measured covariates.
Assume that the first component of \eqn{Z} is 1, and let \eqn{X = (G^T, Z^T)^T}. The
following models are considered:
\deqn{S = \alpha'X + \epsilonS,}
\deqn{Y = \gammaS + \beta'Z + \epsilonY,}
where \eqn{\alpha} and \eqn{\beta} are vectors of regression parameters, \eqn{\gamma}
represents the causal effect of the exposure on the outcome, and
\eqn{(\epsilonS, \epsilonY)^T} is a bivariate normal random vector with mean zero and
an unstructured covariance matrix. We allow a nonzero correlation between \eqn{\epsilonS}
and \eqn{\epsilonY} to account for the potential unmeasured confounders of the
exposure-outcome relationship.

For the argument \code{pheno.dat}, the i-th element of the second column indicates whether the
exposure of the i-th subject is observed, beyond detection limits, or unmeasured, and the
corresponding value in the third column should be the exact measurement on the exposure, the
detection limit, and a dummy value of -999, respectively. Below is a table showing how
the second and the third column should be specified when the exposure variable is observed, beyond
detection limits, or unmeasured.\tabular{lcc}{
    \tab The 2nd column \tab The 3rd column \cr
   Measured and detectable \tab 0 \tab the exact measurement \cr
   Measured but below the lower detection limit \tab 1 \tab the lower detection limit \cr
   Measured but above the upper detection limit \tab 2 \tab the upper detection limit \cr
   Unmeasured \tab 3 \tab -999 \cr
}
}
\references{
Little, R. J., & Rubin, D. B. (2019). Statistical analysis with missing data (3rd Edition). John Wiley & Sons.
}
